refactor(junior): migrate challenge component tree to gjs
