--- junior/app/components/challenge/challenge-actions/component.js
@@ -1,7 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class ChallengeActions extends Component {
-  get hideSkipButton() {
-    return this.args.level === 'TUTORIAL' || this.args.answerHasBeenValidated;
-  }
-}
