--- junior/app/components/challenge/challenge-actions/template.hbs
@@ -1,45 +1,0 @@
-<div class="challenge-actions">
-  {{#unless (or this.hideSkipButton @isLesson)}}
-    <PixButton class="pix1d-button pix1d-button--skip" @triggerAction={{@skipChallenge}} @shape="rounded" @size="large">
-      <span>
-        {{t "pages.challenge.actions.skip"}}
-      </span>
-    </PixButton>
-  {{/unless}}
-  {{#if @answerHasBeenValidated}}
-    <PixButton
-      class="pix1d-button pix1d-button--success"
-      @iconAfter="arrowRight"
-      @triggerAction={{@nextAction}}
-      @size="large"
-    >
-      <span>
-        {{t "pages.challenge.actions.continue"}}
-      </span>
-    </PixButton>
-  {{else if @isLesson}}
-    <PixButton
-      class="pix1d-button"
-      @iconAfter="arrowRight"
-      @isDisabled={{@disableLessonButton}}
-      @triggerAction={{@skipChallenge}}
-      @size="large"
-    >
-      <span>
-        {{t "pages.challenge.actions.continue"}}
-      </span>
-    </PixButton>
-  {{else}}
-    <PixButton
-      class="pix1d-button"
-      @isDisabled={{@disableCheckButton}}
-      @triggerAction={{@validateAnswer}}
-      @size="large"
-    >
-      <span>
-        {{t "pages.challenge.actions.check"}}
-      </span>
-    </PixButton>
-  {{/if}}
-
-</div>
