--- junior/app/components/challenge/challenge-embed-simulator/template.hbs
@@ -1,29 +1,0 @@
-{{! template-lint-disable style-concatenation no-inline-styles }}
-<div class="challenge-embed-simulator">
-  {{#if @hideSimulator}}
-    <div class="challenge-embed-simulator__overlay"></div>
-  {{/if}}
-
-  <iframe
-    tabindex={{if @hideSimulator "-1"}}
-    class="challenge-embed-simulator__iframe"
-    src="{{@url}}"
-    title="{{@title}}"
-    style="{{this.embedDocumentHeightStyle}}"
-    {{did-render this.setIframeHtmlElement}}
-  >
-  </iframe>
-  {{#if @shouldDisplayRebootButton}}
-    <div class="reboot-container">
-      <PixButton
-        class="link link--grey reboot-container__content"
-        aria-label={{t "pages.challenge.embed-simulator.actions.reset-label"}}
-        @iconBefore="refresh"
-        @variant="tertiary"
-        @triggerAction={{this.rebootSimulator}}
-      >
-        {{t "components.challenge.embed-simulator.actions.reset"}}
-      </PixButton>
-    </div>
-  {{/if}}
-</div>
