+++ junior/app/components/challenge/challenge-item-auto-reply.js
