--- junior/app/components/challenge/challenge-item-qcm/template.hbs
@@ -1,17 +1,0 @@
-<div class="challenge-item-proposals__qcm-checkboxes">
-  <p class="challenge-item-proposals__qcm-checkboxes__hint">{{t "pages.challenge.qcm-hint"}}</p>
-  {{#each this.labeledCheckboxes as |labeledCheckbox|}}
-    <PixCheckbox
-      name="{{labeledCheckbox.value}}"
-      @class="pix1d-checkbox {{if @isDisabled 'pix1d-checkbox--disabled'}}"
-      disabled={{@isDisabled}}
-      checked={{labeledCheckbox.checked}}
-      {{on "click" (fn this.checkboxClicked labeledCheckbox.value)}}
-      data-test="challenge-response-proposal-selector"
-    >
-      <:label>
-        <MarkdownToHtml @class="proposal-text" @markdown={{labeledCheckbox.label}} />
-      </:label>
-    </PixCheckbox>
-  {{/each}}
-</div>
