--- junior/app/components/challenge/challenge-item-qcu/template.hbs
@@ -1,19 +1,0 @@
-<div class="challenge-item-proposals__qcu-radios">
-  <p class="challenge-item-proposals__qcu-radios__hint">{{t "pages.challenge.qcu-hint"}}</p>
-  {{#each this.labeledRadios as |labeledRadio|}}
-    <PixRadioButton
-      name="radio"
-      @value={{labeledRadio.value}}
-      disabled={{@isDisabled}}
-      checked={{labeledRadio.checked}}
-      {{on "click" (fn this.radioClicked labeledRadio.value)}}
-      data-value="{{labeledRadio.value}}"
-      @class="pix1d-radio {{if @isDisabled 'pix1d-radio--disabled'}}"
-      data-test="challenge-response-proposal-selector"
-    >
-      <:label>
-        <MarkdownToHtml @class="qcu-panel__text proposal-text" @markdown={{labeledRadio.label}} />
-      </:label>
-    </PixRadioButton>
-  {{/each}}
-</div>
