--- junior/app/components/challenge/challenge-item-qrocm/template.hbs
@@ -1,100 +1,0 @@
-<div {{did-render this.resetAnswerValues}} class="challenge-item-proposals">
-  {{#each this.proposalBlocks as |block|}}
-    {{#if block.showText}}
-      <MarkdownToHtml @markdown={{block.text}} @extensions="remove-paragraph-tags" class="qrocm-proposal__label" />
-    {{/if}}
-    {{#if (eq block.type "select")}}
-      <PixSelect
-        class="challenge-item-proposals__response challenge-item-proposals__response--selector"
-        @value={{get this.answerValues block.input}}
-        @screenReaderOnly={{true}}
-        @placeholder={{block.placeholder}}
-        @hideDefaultOption={{true}}
-        @options={{block.options}}
-        @onChange={{fn this.onSelectChange block.input}}
-        @id="{{block.input}}"
-        @isDisabled={{@isDisabled}}
-      >
-        <:label>{{block.ariaLabel}}</:label>
-      </PixSelect>
-    {{else if (eq block.type "input")}}
-      {{#if block.input}}
-        <div class="qrocm-proposal__input">
-          {{#if block.text}}
-            <label for="{{block.input}}">
-              <MarkdownToHtml
-                @markdown={{block.text}}
-                @extensions="remove-paragraph-tags"
-                class="challenge-item-proposals__input-label"
-              />
-            </label>
-          {{/if}}
-
-          {{#if (eq @challenge.format "paragraphe")}}
-            <PixTextarea
-              @id="{{block.input}}"
-              class="challenge-item-proposals__response challenge-item-proposals__response--paragraph"
-              rows="5"
-              aria-label={{block.ariaLabel}}
-              placeholder={{block.placeholder}}
-              name={{block.randomName}}
-              autocomplete="nope"
-              disabled={{@isDisabled}}
-              {{! To activate validation button as soon as possible }}
-              {{on "keyup" (fn this.onInputChange block.input)}}
-              {{! To also deal with browser dictionary auto complete }}
-              {{on "change" (fn this.onInputChange block.input)}}
-            />
-          {{else if (eq @challenge.format "phrase")}}
-            <PixInput
-              class="challenge-item-proposals__response"
-              @type="text"
-              @id="{{block.input}}"
-              name={{block.randomName}}
-              autocomplete="nope"
-              placeholder={{block.placeholder}}
-              aria-label={{block.ariaLabel}}
-              disabled={{@isDisabled}}
-              {{! To activate validation button as soon as possible }}
-              {{on "keyup" (fn this.onInputChange block.input)}}
-              {{! To also deal with browser dictionary auto complete }}
-              {{on "change" (fn this.onInputChange block.input)}}
-            />
-          {{else if (eq @challenge.format "nombre")}}
-            <PixInput
-              @id="{{block.input}}"
-              class="challenge-item-proposals__response challenge-item-proposals__response--number"
-              name={{block.randomName}}
-              type="number"
-              min="0"
-              placeholder={{block.placeholder}}
-              @value={{get @answerValues block.input}}
-              aria-label={{block.ariaLabel}}
-              disabled={{@isDisabled}}
-              {{! To activate validation button as soon as possible }}
-              {{on "keyup" (fn this.onInputChange block.input)}}
-              {{! To also deal with browser dictionary auto complete }}
-              {{on "change" (fn this.onInputChange block.input)}}
-            />
-          {{else}}
-            <PixInput
-              @type="text"
-              class="challenge-item-proposals__response challenge-item-proposals__response--short-text"
-              @id="{{block.input}}"
-              name={{block.randomName}}
-              autocomplete="nope"
-              placeholder={{block.placeholder}}
-              @value={{get @answerValues block.input}}
-              aria-label={{block.ariaLabel}}
-              disabled={{@isDisabled}}
-              {{! To activate validation button as soon as possible }}
-              {{on "keyup" (fn this.onInputChange block.input)}}
-              {{! To also deal with browser dictionary auto complete }}
-              {{on "change" (fn this.onInputChange block.input)}}
-            />
-          {{/if}}
-        </div>
-      {{/if}}
-    {{/if}}
-  {{/each}}
-</div>
