--- junior/app/components/challenge/challenge-web-component/template.hbs
@@ -1,3 +1,0 @@
-{{#if (eq @tagName "qcu-image")}}
-  <qcu-image props={{@props}} {{on "answer" this.handleAnswer}} />
-{{/if}}
