--- junior/app/components/challenge/item/template.hbs
@@ -1,81 +1,0 @@
-<div class="challenge-item {{unless this.isMediaWithForm 'challenge-item--single-display'}}">
-  {{#if this.hasMedia}}
-    <div class="challenge-item__media {{unless this.isMediaWithForm 'challenge-item__media--single-display'}}">
-      {{#if @challenge.illustrationUrl}}
-        <div class="challenge-item__image">
-          <Challenge::ChallengeMedia @src={{@challenge.illustrationUrl}} @alt={{@challenge.illustrationAlt}} />
-        </div>
-      {{/if}}
-      {{#if @challenge.hasValidEmbedDocument}}
-        <div class="challenge-item__embed">
-          <Challenge::ChallengeEmbedSimulator
-            @url={{@challenge.embedUrl}}
-            @title={{@challenge.embedTitle}}
-            @height={{@challenge.embedHeight}}
-            @hideSimulator={{@isDisabled}}
-            @isMediaWithForm={{this.isMediaWithForm}}
-            @shouldDisplayRebootButton={{this.shouldDisplayRebootButton}}
-          />
-        </div>
-      {{/if}}
-      {{#if @challenge.hasWebComponent}}
-        <div class="challenge-item__web-component">
-          <Challenge::ChallengeWebComponent
-            @tagName={{@challenge.webComponentTagName}}
-            @props={{@challenge.webComponentProps}}
-            @setAnswerValue={{@setAnswerValue}}
-          />
-        </div>
-      {{/if}}
-    </div>
-  {{/if}}
-  <div class="challenge-item__proposals {{unless this.isMediaWithForm 'challenge-item__proposals--single-display'}}">
-    {{#if @challenge.autoReply}}
-      <div class="challenge-item__autoreply">
-        <Challenge::ChallengeItemAutoReply @setAnswerValue={{@setAnswerValue}} />
-      </div>
-    {{/if}}
-    {{#if (or @challenge.isQROC @challenge.isQROCM)}}
-      <div class="challenge-item__qrocm">
-        <Challenge::ChallengeItemQrocm
-          @challenge={{@challenge}}
-          @setAnswerValue={{@setAnswerValue}}
-          @isDisabled={{@isDisabled}}
-        />
-      </div>
-    {{/if}}
-    {{#if @challenge.isQCU}}
-      <div class="challenge-item__qcu">
-        <Challenge::ChallengeItemQcu
-          @challenge={{@challenge}}
-          @setAnswerValue={{@setAnswerValue}}
-          @assessment={{@assessment}}
-          @isDisabled={{@isDisabled}}
-        />
-      </div>
-    {{/if}}
-    {{#if @challenge.isQCM}}
-      <div class="challenge-item__qcm">
-        <Challenge::ChallengeItemQcm
-          @challenge={{@challenge}}
-          @setAnswerValue={{@setAnswerValue}}
-          @setValidationWarning={{@setValidationWarning}}
-          @assessment={{@assessment}}
-          @isDisabled={{@isDisabled}}
-        />
-      </div>
-    {{/if}}
-    <div class="container__actions">
-      <Challenge::ChallengeActions
-        @validateAnswer={{@validateAnswer}}
-        @skipChallenge={{@skipChallenge}}
-        @level={{@activity.level}}
-        @nextAction={{@resume}}
-        @isLesson={{@challenge.focused}}
-        @disableCheckButton={{@disableCheckButton}}
-        @disableLessonButton={{@disableLessonButton}}
-        @answerHasBeenValidated={{@answerHasBeenValidated}}
-      />
-    </div>
-  </div>
-</div>
