+++ junior/app/templates/assessment/challenge.hbs
@@ -1,4 +1,4 @@
-<Challenge
+<Challenge::Challenge
   @challenge={{@model.challenge}}
   @assessment={{@model.assessment}}
   @activity={{@model.activity}}
