+++ junior/app/templates/challenge-preview.hbs
@@ -1,1 +1,1 @@
-<Challenge @challenge={{@model}} />
+<Challenge::Challenge @challenge={{@model}} />
