+++ junior/tests/acceptance/challenge-item-qroc-test.js
@@ -17,6 +17,7 @@ module('Acceptance | Displaying a QROC challenge', function (hooks) {
     test('should render challenge information and question', async function (assert) {
       // when
       const screen = await visit(`/assessments/${assessment.id}/challenges`);
+      debugger;
       // then
       assert.dom('.challenge-item-proposals__response').exists({ count: 1 });
       assert.dom(screen.getByText('Rue de :')).exists();
