+++ junior/tests/helpers/index.js
@@ -1,6 +1,10 @@
 import { setupMirage } from 'ember-cli-mirage/test-support';
 import { setupIntl, t } from 'ember-intl/test-support';
-import { setupApplicationTest as upstreamSetupApplicationTest, setupTest as upstreamSetupTest } from 'ember-qunit';
+import {
+  setupApplicationTest as upstreamSetupApplicationTest,
+  setupRenderingTest,
+  setupTest as upstreamSetupTest,
+} from 'ember-qunit';
 
 // This file exists to provide wrappers around ember-qunit's
 // test setup functions. This way, you can easily extend the setup that is
@@ -31,4 +35,9 @@ function setupTest(hooks, options) {
   // Additional setup for unit tests can be done here.
 }
 
-export { setupApplicationTest, setupTest, t };
+export default function setupIntlRenderingTest(hooks) {
+  setupRenderingTest(hooks);
+  setupIntl(hooks, 'fr');
+}
+
+export { setupApplicationTest, setupIntlRenderingTest, setupRenderingTest, setupTest, t };
