+++ junior/tests/integration/bubble_test.gjs
@@ -1,7 +1,7 @@
-import { render } from '@1024pix/ember-testing-library';
+import {render} from '@1024pix/ember-testing-library';
 import Bubble from 'junior/components/bubble';
-import { setupRenderingTest } from 'junior/helpers/tests';
-import { module, test } from 'qunit';
+import {setupRenderingTest} from '../helpers';
+import {module, test} from 'qunit';
 
 module('Integration | Component | Bubble', function (hooks) {
   setupRenderingTest(hooks);
